requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
